Add files via upload
This is a web-based "Guess The Number" game where the player tries to guess a randomly generated secret number between 1 and 100. The game provides hints (higher/lower) and indicates when the guess is very close to the secret number. It also tracks the number of attempts and allows the user to play again.